requirements.txt
Django==4.2.3
djangorestframework==3.14.0
drf-spectacular==0.24.2
psycopg2-binary==2.9.6
model-bakery==1.13.0
pillow==9.2.0
python-magic==0.4.27